implemented typing Self in series.py and moved date range adjusting methods to CommonModel
