+++ openseries/series.py
@@ -24,17 +24,13 @@ from pandas import (
     date_range,
 )
 from pydantic import model_validator
+from typing_extensions import Self
 
 from openseries._common_model import _CommonModel
 from openseries._risk import (
     _ewma_calc,
 )
-from openseries.datefixer import (
-    _get_calc_range,
-    align_dataframe_to_local_cdays,
-    date_fix,
-    do_resample_to_business_period_ends,
-)
+from openseries.datefixer import date_fix, do_resample_to_business_period_ends
 from openseries.types import (
     Countries,
     CountriesType,
@@ -110,7 +106,7 @@ class OpenTimeSeries(_CommonModel):
     label: Optional[str] = None
 
     @model_validator(mode="after")  # type: ignore[misc,unused-ignore]
-    def dates_and_values_validate(self: OpenTimeSeries) -> OpenTimeSeries:
+    def dates_and_values_validate(self: Self) -> Self:
         """Pydantic validator to ensure dates and values are validated."""
         values_list_length = len(self.values)
         dates_list_length = len(self.dates)
@@ -377,7 +373,7 @@ class OpenTimeSeries(_CommonModel):
             ),
         )
 
-    def from_deepcopy(self: OpenTimeSeries) -> OpenTimeSeries:
+    def from_deepcopy(self: Self) -> Self:
         """
         Create copy of OpenTimeSeries object.
 
@@ -388,7 +384,7 @@ class OpenTimeSeries(_CommonModel):
         """
         return deepcopy(self)
 
-    def pandas_df(self: OpenTimeSeries) -> OpenTimeSeries:
+    def pandas_df(self: Self) -> Self:
         """
         Populate .tsdf Pandas DataFrame from the .dates and .values lists.
 
@@ -407,54 +403,8 @@ class OpenTimeSeries(_CommonModel):
 
         return self
 
-    def calc_range(
-        self: OpenTimeSeries,
-        months_offset: Optional[int] = None,
-        from_dt: Optional[dt.date] = None,
-        to_dt: Optional[dt.date] = None,
-    ) -> tuple[dt.date, dt.date]:
-        """
-        Create user defined date range.
-
-        Parameters
-        ----------
-        months_offset: int, optional
-            Number of months offset as positive integer. Overrides use of from_date
-            and to_date
-        from_dt: datetime.date, optional
-            Specific from date
-        to_dt: datetime.date, optional
-            Specific from date
-
-        Returns
-        -------
-        (datetime.date, datetime.date)
-            Start and end date of the chosen date range
-        """
-        return _get_calc_range(
-            data=self.tsdf,
-            months_offset=months_offset,
-            from_dt=from_dt,
-            to_dt=to_dt,
-        )
-
-    def align_index_to_local_cdays(self: OpenTimeSeries) -> OpenTimeSeries:
-        """
-        Align the index .tsdf with local calendar business days.
-
-        Returns
-        -------
-        OpenTimeSeries
-            An OpenTimeSeries object
-        """
-        self.tsdf = align_dataframe_to_local_cdays(
-            data=self.tsdf,
-            countries=self.countries,
-        )
-        return self
-
     def all_properties(
-        self: OpenTimeSeries,
+        self: Self,
         properties: Optional[list[LiteralSeriesProps]] = None,
     ) -> DataFrame:
         """
@@ -482,7 +432,7 @@ class OpenTimeSeries(_CommonModel):
         return pdf
 
     @property
-    def worst_month(self: OpenTimeSeries) -> float:
+    def worst_month(self: Self) -> float:
         """
         Most negative month.
 
@@ -502,7 +452,7 @@ class OpenTimeSeries(_CommonModel):
             ).iloc[0],
         )
 
-    def value_to_ret(self: OpenTimeSeries) -> OpenTimeSeries:
+    def value_to_ret(self: Self) -> Self:
         """
         Convert series of values into series of returns.
 
@@ -522,10 +472,7 @@ class OpenTimeSeries(_CommonModel):
         )
         return self
 
-    def value_to_diff(
-        self: OpenTimeSeries,
-        periods: int = 1,
-    ) -> OpenTimeSeries:
+    def value_to_diff(self: Self, periods: int = 1) -> Self:
         """
         Convert series of values to series of their period differences.
 
@@ -551,7 +498,7 @@ class OpenTimeSeries(_CommonModel):
         )
         return self
 
-    def to_cumret(self: OpenTimeSeries) -> OpenTimeSeries:
+    def to_cumret(self: Self) -> Self:
         """
         Convert series of returns into cumulative series of values.
 
@@ -578,10 +525,10 @@ class OpenTimeSeries(_CommonModel):
         return self
 
     def from_1d_rate_to_cumret(
-        self: OpenTimeSeries,
+        self: Self,
         days_in_year: int = 365,
         divider: float = 1.0,
-    ) -> OpenTimeSeries:
+    ) -> Self:
         """
         Convert series of 1-day rates into series of cumulative values.
 
@@ -615,9 +562,9 @@ class OpenTimeSeries(_CommonModel):
         return self
 
     def resample(
-        self: OpenTimeSeries,
+        self: Self,
         freq: Union[LiteralBizDayFreq, str] = "BM",
-    ) -> OpenTimeSeries:
+    ) -> Self:
         """
         Resamples the timeseries frequency.
 
@@ -638,11 +585,11 @@ class OpenTimeSeries(_CommonModel):
         return self
 
     def resample_to_business_period_ends(
-        self: OpenTimeSeries,
+        self: Self,
         freq: LiteralBizDayFreq = "BM",
         convention: LiteralPandasResampleConvention = "end",
         method: LiteralPandasReindexMethod = "nearest",
-    ) -> OpenTimeSeries:
+    ) -> Self:
         """
         Resamples timeseries frequency to the business calendar month end dates.
 
@@ -676,7 +623,7 @@ class OpenTimeSeries(_CommonModel):
         return self
 
     def ewma_vol_func(
-        self: OpenTimeSeries,
+        self: Self,
         lmbda: float = 0.94,
         day_chunk: int = 11,
         dlta_degr_freedms: int = 0,
@@ -758,10 +705,10 @@ class OpenTimeSeries(_CommonModel):
         )
 
     def running_adjustment(
-        self: OpenTimeSeries,
+        self: Self,
         adjustment: float,
         days_in_year: int = 365,
-    ) -> OpenTimeSeries:
+    ) -> Self:
         """
         Add or subtract a fee from the timeseries return.
 
@@ -820,12 +767,12 @@ class OpenTimeSeries(_CommonModel):
         return self
 
     def set_new_label(
-        self: OpenTimeSeries,
+        self: Self,
         lvl_zero: Optional[str] = None,
         lvl_one: Optional[ValueType] = None,
         *,
         delete_lvl_one: bool = False,
-    ) -> OpenTimeSeries:
+    ) -> Self:
         """
         Set the column labels of the .tsdf Pandas Dataframe.
